requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
PyMySQL==1.1.0
Werkzeug==2.3.7

